Ensure that the code from head.o is put first in the output
The output was broken when fc519387 changed the order in which the
object files are passed to the linker. Rather than controlling this
from the Makefile, we change the linker script to be explicit about
which object we want to go first.

Note that the ENTRY() statement correctly sets the entry point in
the ELF file, but has no effect on what code is placed first, so as
soon as we convert from ELF to unstructured binary the information
about the entry point is lost.