Rename src/head.S to src/head-a320.S and target.ld to target-a320.ld
'head.S' cannot contain platform-independent code; the jz4740 and
the jz4770 don't expect the same value as the first dword of the
binary (jz4740 expects 0xFF as the first byte while jz4770 expects
'MSPL' as the first dword).

'target.ld' cannot be used for all platforms because the RAM start
address of the program is different among them.