Copy boot loader to SDRAM once that is up and running
Tests on JZ4770 show that the data cache lines that contain the boot
loader are not marked as dirty initially. Therefore, if those cache
lines are evicted, the data is lost. To avoid that, we copy the data
into SDRAM, by copying from kseg0 (cached) to the same area in kseg1
(uncached).