Add workaround to boot older GCW Zero kernels
Older GCW Zero kernels contain mininit 1.x in their initramfs and have
a built-in kernel command line that is appended to the parameters
passed to the kernel by UBIBoot. Newer GCW Zero kernels contain
neither an initramfs nor a built-in kernel command line and are
instructed by UBIBoot to load mininit 2.x from the system partition.

If we run mininit 1.x, the parameters set by UBIBoot break mininit's
mounting of the rootfs: UBIBoot makes the "root=" parameter point to
the system partition instead of "/dev/loop0".

If we run mininit 2.x (using "rdinit=/skip"), the parameters from the
built-in kernel command line break the kernel's mounting of the rootfs:
the kernel's built-in "root=" parameter points to "/dev/loop0" instead
of the system partition.

The workaround relies on differences in parameter parsing between the
kernel (last occurrence counts) and mininit 1.x (first occurrence
counts). It's ugly, but the alternatives seem worse. The workaround is
guarded by the MININIT1_COMPAT define, so for any new board the
workaround can be skipped simply by not defining that the board config.