Spread body of mmc_block_read() over start/receive/stop functions
This is a preparation to allow data to be processed while the read
command is in progress.

This commit changes the handling of read errors: when a block read
fails, a stop command is sent. I think this will have a better
chance of leaving the system in a state where subsequent load
attempts can succeed.