Move code to end of address space to allow stack to grow
As the code size grew, the stack could overwrite global variables.
By moving the code behind the global variables, the stack can grow
much larger. Note that the stack overlapping the code (.text) section
is not a problem because the code is read from the instruction cache
and the stack is written into the data cache.