Cut down memmove amount significantly when loading a uImage
Previously the entire image would be loaded at the default load
address and the the body (kernel) would be moved to the right location
after loading. Now we only load the first sector at the default load
address, then examine the header, move the remainder of the data and
continue loading at the right location. This saves over 300 ms.