Do not make copies of constant data to the stack
In C, declarations like the following:

  int data[] = { 1, 2, 3 };

create a new modifiable copy of the array on the stack. Since the JZ
register manipulation functions never modify these copies, they were
better declared as static const. This makes all of the code refer to
a single copy in the data segment instead.