UBI: Change algorithm
Instead of parsing the NAND twice and pre-caching all the eraseblocks
corresponding to all the UBI volumes found on the flash, which takes
a lot of our precious dcache, we now parse the NAND three times:

Step #1: Look for the eraseblock containing the volume table. It should
         be one of the first eraseblocks of the partition, so it does
	 not take much time.

Step #2: Get the address of all the eraseblocks that correspond to our
         UBI volume.

Step #3: Load the data from the eraseblocks in the correct order.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>