UBI: Reintroduce the volume table parser
Recently, the code parsing the volume names was dropped with the
assumption that the regular kernel will always be the volume 0, and the
backup kernel will always be the volume 1.

This did not work at all with the firmware update procedure, which
atomically swaps the volume numbers between the regular and backup
kernels.

This commit modifies the NAND parsing loop so that it will populate
three linked lists, one for the volume table, one for the main kernel,
one for the backup kernel, in a single loop. Then, the ID of the
requested kernel will be loaded from the volume table and the EBs we
pre-populated will be loaded to RAM.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>