lepus: Set PLL to 1056 MHz on JZ4760, 1200 MHz on JZ4760B
Previous commit changed the PLL from 528 MHz to 1200 MHz, and doubled
the dividers, as it gave a better performance.

However, this caused the CPU clock to bump from 528 MHz to 600 MHz,
which is out-of-spec for the JZ4760.

Lower this back to 528 MHz by setting the PLL to 1056 MHz when a JZ4760
is detected, or set it to 1200 MHz when a JZ4760B is detected.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>