rs90: Add Alt3 button for RG-99
The RS-90 has only two buttons that can be used in the bootloader: START
and R2. Others have either an action set in the boot ROM, or aren't
mapped to GPIOs.

Use pin D24 as the "hide boot logo" function. This pin corresponds to
the A button on the RG-99 handheld.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>