Support RG300 with Y5C memory
Theoretically CL=3 at 200 MHz would be 15ns CAS latency and Y5C is
documented for CL=5 and 3ns -> 15 ns. In practice my unit that was
shipped with Y5C memory needs CL=4. Revert to CL=4 by default that
should work for everyone and keep CL=3 as the value for S6C