Remove MBR_PRELOAD_ADDR
Rely on the logic that UBIBoot loaded into RAM with MBR is broken.
Because of full UBiboot might be loaded over USB now - jzboot
patched to do so.

Another issue introduced by MBR_PRELOAD_ADDR is that it prevents
to load real MBR for external MMC (e.g. MMC2 on lepus) and uses same
preloaded MBR for it.

Signed-off-by: Siarhei Volkau <lis8215@gmail.com>